refactor(agent_utils.py): remove redundant model detection functions and move them to model_detection.py for better organization feat(model_detection.py): add model_name_has_claude and is_anthropic_claude functions to enhance model detection capabilities
